Added more quickcheck tests for the lexer
